Moved graph implementation to separate module
